+++ src/main/java/view/MenuOption.java
@@ -1,6 +1,6 @@
 package view;
 
-import constant.ErrorMessage;
+import util.ErrorMessage;
 
 import java.util.Arrays;
