+++ src/test/java/AllCrewTest.java
@@ -1,4 +1,4 @@
-import constant.ErrorMessage;
+import util.ErrorMessage;
 import domain.AllCrew;
 import domain.Attendance;
 import domain.Crew;
